+++ src/utils/completeOptions.test.ts
@@ -97,6 +97,7 @@ describe('completeOptions', () => {
         excludeRegex: null,
         sessionsPercentageKept: 100,
         rejectSession: DEFAULT_SESSION_REJECTION,
+        disableVideoRecording: true,
         window,
       }
       expect(completed).toStrictEqual(expected)
@@ -109,6 +110,7 @@ describe('completeOptions', () => {
         maxDelay: 15,
         sessionsPercentageKept: 33.3,
         rejectSession: customSessionRejection,
+        disableVideoRecording: false,
       })
       const expected: CollectorOptions = {
         authKey: '',
@@ -119,6 +121,7 @@ describe('completeOptions', () => {
         excludeRegex: null,
         sessionsPercentageKept: 33.3,
         rejectSession: customSessionRejection,
+        disableVideoRecording: false,
         window,
       }
       expect(completed).toStrictEqual(expected)
@@ -145,6 +148,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          disableVideoRecording: true,
           window,
         }
         expect(completed).toStrictEqual(expected)
@@ -165,6 +169,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          disableVideoRecording: true,
           window,
         }
         expect(completed).toStrictEqual(expected)
@@ -185,6 +190,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          disableVideoRecording: true,
           window,
         }
         expect(completed).toStrictEqual(expected)
@@ -205,6 +211,7 @@ describe('completeOptions', () => {
           excludeRegex: /^#my-id-.*$/,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          disableVideoRecording: true,
           window,
         }
         expect(completed).toStrictEqual(expected)
@@ -225,6 +232,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          disableVideoRecording: true,
           window,
         }
         expect(completed).toStrictEqual(expected)